requirements.txt
streamlit
mediapipe
opencv-python
numpy
pillow
ipykernel
jupyter
streamlit-webrtc
av
